refaktorer accordion elementer til kvittering
