+++ skribenten-web/frontend/src/routes/saksnummer_/$saksId/kvittering/-components/KvitterteBrev.tsx
@@ -1,12 +1,117 @@
 import { css } from "@emotion/react";
 import { XMarkOctagonFillIcon } from "@navikt/aksel-icons";
-import { Accordion, BodyShort, Button, VStack } from "@navikt/ds-react";
+import { Accordion, BodyShort, Button, Label, Tag, VStack } from "@navikt/ds-react";
+import { useMutation } from "@tanstack/react-query";
 
+import { sendBrev } from "~/api/sak-api-endpoints";
 import { ApiError } from "~/components/ApiError";
+import type { BestillBrevResponse } from "~/types/brev";
 import type { Nullable } from "~/types/Nullable";
 
-import type { FerdigstillResponse } from "./FerdigstillResultatContext";
+import type { FerdigstillResponse, FerdigstillResponser } from "./FerdigstillResultatContext";
 import Oppsummeringspar from "./Oppsummeringspar";
+
+const KvitterteBrev = (properties: { sakId: string; resultat: FerdigstillResponser }) => {
+  return (
+    <Accordion>
+      {properties.resultat.map((result, index) => (
+        <KvittertBrev key={`resultat-${index}`} resultat={result} sakId={properties.sakId} />
+      ))}
+    </Accordion>
+  );
+};
+
+export default KvitterteBrev;
+
+const KvittertBrev = (properties: { sakId: string; resultat: FerdigstillResponse }) => {
+  const mutation = useMutation<BestillBrevResponse, Error, number>({
+    mutationFn: (brevId) => sendBrev(properties.sakId, brevId),
+  });
+
+  return (
+    <Accordion.Item>
+      <ResultatAccordionHeader
+        resultat={
+          mutation.isSuccess
+            ? {
+                status: "fulfilledWithSuccess",
+                brevInfo: properties.resultat.brevInfo,
+                response: mutation.data,
+              }
+            : properties.resultat
+        }
+        sakId={properties.sakId}
+      />
+      <ResultatAccordionContentWrapper
+        error={mutation.isError ? mutation.error : null}
+        isPending={mutation.isPending}
+        onPrøvIgjenClick={() => mutation.mutate(properties.resultat.brevInfo.id)}
+        resultat={
+          mutation.isSuccess
+            ? {
+                status: "fulfilledWithSuccess",
+                brevInfo: properties.resultat.brevInfo,
+                response: mutation.data,
+              }
+            : properties.resultat
+        }
+        sakId={properties.sakId}
+      />
+    </Accordion.Item>
+  );
+};
+
+const ResultatAccordionHeader = (properties: { sakId: string; resultat: FerdigstillResponse }) => {
+  const { tag, tittel } = hentTagOgTittelForHeader(properties.resultat);
+
+  return (
+    <Accordion.Header>
+      <div
+        css={css`
+          display: flex;
+          flex-direction: column;
+          align-items: flex-start;
+        `}
+      >
+        {tag}
+        <Label>{tittel}</Label>
+      </div>
+    </Accordion.Header>
+  );
+};
+
+const hentTagOgTittelForHeader = (resultat: FerdigstillResponse) => {
+  switch (resultat.status) {
+    case "fulfilledWithError": {
+      const tag = (
+        <Tag size="small" variant={"error"}>
+          Brev ble ikke sendt
+        </Tag>
+      );
+
+      return { tag, tittel: resultat.brevInfo.brevkode };
+    }
+    case "fulfilledWithSuccess": {
+      const tag = resultat.response.journalpostId ? (
+        /*
+                  if(lokaltprint){
+                    return <Tag variant={"info"} size="small">Sendt til lokalprint</Tag>
+                    } else {
+              */
+        //teknisk sett er den bare journalført
+        <Tag size="small" variant={"success"}>
+          Sendt til mottaker
+        </Tag>
+      ) : (
+        <Tag size="small" variant={"error"}>
+          Brev ble ikke sendt
+        </Tag>
+      );
+
+      return { tag, tittel: resultat.brevInfo.brevkode };
+    }
+  }
+};
 
 const ResultatAccordionContentWrapper = (properties: {
   sakId: string;
@@ -148,5 +253,3 @@ const ResultatContentError = (properties: {
     </VStack>
   );
 };
-
-export default ResultatAccordionContentWrapper;
