+++ skribenten-web/frontend/src/routes/saksnummer_/$saksId/kvittering/route.tsx
@@ -1,15 +1,9 @@
 import { css } from "@emotion/react";
-import { Accordion, Button, Heading, VStack } from "@navikt/ds-react";
-import { useMutation } from "@tanstack/react-query";
+import { Button, Heading, VStack } from "@navikt/ds-react";
 import { createFileRoute, useNavigate } from "@tanstack/react-router";
 
-import { sendBrev } from "~/api/sak-api-endpoints";
-import type { BestillBrevResponse } from "~/types/brev";
-
-import type { FerdigstillResponse, FerdigstillResponser } from "./-components/FerdigstillResultatContext";
 import { useFerdigstillResultatContext } from "./-components/FerdigstillResultatContext";
-import ResultatAccordionContent from "./-components/ResultatAccordionContent";
-import ResultatAccordionHeader from "./-components/ResultatAccordionHeader";
+import KvitterteBrev from "./-components/KvitterteBrev";
 
 export const Route = createFileRoute("/saksnummer/$saksId/kvittering")({
   component: Kvittering,
@@ -94,55 +88,7 @@ function Kvittering() {
           width: 1px;
         `}
       ></div>
-      <FerdigstillingsResultat resultat={ferdigstillBrevContext.resultat} sakId={saksId} />
+      <KvitterteBrev resultat={ferdigstillBrevContext.resultat} sakId={saksId} />
     </div>
   );
 }
-
-const FerdigstillingsResultat = (properties: { sakId: string; resultat: FerdigstillResponser }) => {
-  return (
-    <Accordion>
-      {properties.resultat.map((result, index) => (
-        <ResultatAccordionItem key={`resultat-${index}`} resultat={result} sakId={properties.sakId} />
-      ))}
-    </Accordion>
-  );
-};
-
-const ResultatAccordionItem = (properties: { sakId: string; resultat: FerdigstillResponse }) => {
-  const mutation = useMutation<BestillBrevResponse, Error, number>({
-    mutationFn: (brevId) => sendBrev(properties.sakId, brevId),
-  });
-
-  return (
-    <Accordion.Item>
-      <ResultatAccordionHeader
-        resultat={
-          mutation.isSuccess
-            ? {
-                status: "fulfilledWithSuccess",
-                brevInfo: properties.resultat.brevInfo,
-                response: mutation.data,
-              }
-            : properties.resultat
-        }
-        sakId={properties.sakId}
-      />
-      <ResultatAccordionContent
-        error={mutation.isError ? mutation.error : null}
-        isPending={mutation.isPending}
-        onPrøvIgjenClick={() => mutation.mutate(properties.resultat.brevInfo.id)}
-        resultat={
-          mutation.isSuccess
-            ? {
-                status: "fulfilledWithSuccess",
-                brevInfo: properties.resultat.brevInfo,
-                response: mutation.data,
-              }
-            : properties.resultat
-        }
-        sakId={properties.sakId}
-      />
-    </Accordion.Item>
-  );
-};
